refactor(gateway): extract background watchers to a mixin (mechanical move)
gateway/run.py is 26,877 lines. This continues the decomposition that already
produced GatewayAuthorizationMixin, GatewayKanbanWatchersMixin and
GatewaySlashCommandsMixin, moving the long-running monitor coroutines out to
gateway/watchers.py:

  _redeliver_pending_obligations, _schedule_resume_pending_sessions,
  _session_expiry_watcher, _platform_reconnect_watcher,
  _run_background_task_inner, _watch_update_progress, _run_process_watcher

`_dispose_unused_adapter` moves with them -- the reconnect watcher is its only
caller -- and is re-exported from gateway.run so the existing
`from gateway.run import _dispose_unused_adapter` path keeps working.

Behaviour-neutral by construction: all 8 bodies are byte-identical to their
previous text, verified by an AST comparison that strips exactly the injected
import statements and nothing else. The only edits are deferred
`from gateway.run import ...` lines for the module-level names the bodies still
reach for -- resolved at call time, the pattern the earlier phases established,
so this module never imports gateway.run, nothing circularises, and a test
patching `gateway.run.X` still takes effect. The module binds
logger = logging.getLogger("gateway.run") so extracted log records keep their
original logger name, matching gateway/kanban_watchers.py.

run.py: 26,877 -> 25,567 lines.

Rebuilt on current main rather than rebased: the branch was 1,035 commits
behind and gateway/run.py conflicted. For a mechanical move, re-deriving the
extraction guarantees the moved bodies match today's main byte for byte, which
a hand-resolved conflict cannot.

test_background_and_main_agent_paths_call_refresh counted call sites by reading
gateway/run.py's source text, which AGENTS.md bans and which reports a false
regression the moment one of those sites moves -- the only thing this PR does.
Replaced with behavioural coverage: the background path drives the real
_run_background_task_inner and asserts the chain reaching AIAgent is the freshly
reloaded one rather than the frozen startup snapshot, and the cached-agent path
asserts a refreshed chain lands on a stale agent.
